backend/apps/contact/serializers.py
from rest_framework import serializers

from .models import ContactMessage


class ContactMessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = ContactMessage
        fields = ["id", "name", "email", "subject", "message", "is_processed", "created_at"]
        read_only_fields = ["id", "created_at"]
