backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
django-storages==1.14.4
boto3==1.35.85
celery==5.4.0
redis==5.2.1
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
